Initial migrations and models
To simplify the beginning of the project, there will be 4 simples models: User, Group, Member and Payment.
The User will represent each user, with its name, email and id.
The Group is simply the definition of a group, which can be a shared house, the company expenses or a trip. It will contain only the name and description.
The members table will expose the relation between users and groups, therefore a Member is a User within a Group.
A Payment represents a value a Member paid and its description. Since only Members can spend something, the relation of which group it belongs is within.