Implements flash messasing
In order to implement flash messaging, the first Vue component was with two possible ways to use: sending the message on the session and calling the method flash in a script, that will be used for future ajax request.

To test, it was already implemented it's use on the creation of a group and payments.