Creating Seeders and specific factories
In order to easily populate the database when starting to test the project, I made some seeders, so a new developer can easily test the application. Another improvement of this commit was to separate the factories into specific ones. This way it is easier to know where find the factories.