Creates Form to split payments
Now it is possible to split the payments at the specific group page, by selecting all the desired payments and then selecting the option Split. This will open a modal with a form, which was commited here, and then the request is sent to the controller.

Next I will prepare better feedback for the user and also a place to see from who the user needs to pay or receive.